+++ node/main-globals.ts
@@ -1,4 +1,4 @@
-import { ScreenshotSettings, InputSources } from './interfaces/final-object.interface';
+import { ScreenshotSettings, InputSources } from '../interfaces/final-object.interface';
 
 export const GLOBALS: VhaGlobals = {
   angularApp: null,            // reference used to send messages back to Angular App
