+++ node/main-support.ts
@@ -18,8 +18,8 @@ const ffprobePath = require('@ffprobe-installer/ffprobe').path.replace('app.asar
 
 import { acceptableFiles } from './main-filenames';
 import { GLOBALS, VhaGlobals } from './main-globals';
-import { FinalObject, ImageElement, NewImageElement, ScreenshotSettings, InputSources } from './interfaces/final-object.interface';
-import { ResolutionString } from './src/app/pipes/resolution-filter.service';
+import { FinalObject, ImageElement, NewImageElement, ScreenshotSettings, InputSources } from '../interfaces/final-object.interface';
+import { ResolutionString } from '../src/app/pipes/resolution-filter.service';
 import { Stats } from 'fs';
 import { queueThumbExtraction } from './main-extract';
