+++ src/app/components/wizard/wizard.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 import { AllowedScreenshotHeight } from '../../../../interfaces/final-object.interface';
 import { HistoryItem } from '../../../../interfaces/shared-interfaces';
-import { ImportStage } from '../../../../main-support';
+import { ImportStage } from '../../../../node/main-support';
 import { WizardOptions } from '../../../../interfaces/wizard-options.interface';
 
 import { historyItemRemove, slowFadeIn } from '../../common/animations';
